+++ Libplanet/Store/IStore.cs
@@ -232,6 +232,42 @@ namespace Libplanet.Store
         TxExecution GetTxExecution(BlockHash blockHash, TxId txid);
 
         /// <summary>
+        /// Records a index for given pair <paramref name="txId"/> and <paramref name="blockHash"/>.
+        /// If there exist a record for <paramref name="txId"/> already,
+        /// it overwrites the record silently.
+        /// </summary>
+        /// <param name="chainId">The chain ID of the index.</param>
+        /// <param name="txId">The <see cref="TxId"/> of the <see cref="Transaction{T}"/>.</param>
+        /// <param name="blockHash">The <see cref="BlockHash"/> of the <see cref="Block{T}"/>.
+        /// </param>
+        void PutTxIdBlockHashIndex(Guid chainId, TxId txId, BlockHash blockHash);
+
+        /// <summary>
+        /// Determines whether the <see cref="IStore"/> contains index to <see cref="BlockHash"/>
+        /// with given <paramref name="txId"/>.
+        /// </summary>
+        /// <param name="chainId">The chain ID of the index.</param>
+        /// <param name="txId">The <see cref="TxId"/> of the <see cref="Transaction{T}"/>.</param>
+        /// <returns><c>true</c> if the index with given <paramref name="txId"/> is exists.
+        /// Otherwise, <c>false</c>.</returns>
+        bool HasTxIdBlockHashIndex(Guid chainId, TxId txId);
+
+        /// <summary>
+        /// Retrieves the <see cref="BlockHash"/> indexed by the <paramref name="txId"/>.
+        /// </summary>
+        /// <param name="chainId">The chain ID of the index.</param>
+        /// <param name="txId">The <see cref="TxId"/> of the <see cref="Transaction{T}"/>.</param>
+        /// <returns><see cref="BlockHash"/> if the index exists. Otherwise <c>null</c>.</returns>
+        BlockHash? GetTxIdBlockHashIndex(Guid chainId, TxId txId);
+
+        /// <summary>
+        /// Deletes the index for the <paramref name="txId"/>.
+        /// </summary>
+        /// <param name="chainId">The chain ID of the index.</param>
+        /// <param name="txId">The <see cref="TxId"/> of the <see cref="Transaction{T}"/>.</param>
+        void DeleteTxIdBlockHashIndex(Guid chainId, TxId txId);
+
+        /// <summary>
         /// Records the perceived time of a block.  If there is already a record, it is overwritten.
         /// </summary>
         /// <param name="blockHash"><see cref="Block{T}.Hash"/> to record its perceived time.