+++ CHANGES.md
@@ -153,6 +153,14 @@ To be released.
      -  Removed `Transaction<T>.EvaluateActions()` method.
  -  Parameters `action`, `inputContext`, and `outputStates` for
     `ActionEvaluation()` constructor can no longer be `null`.  [[#1305]]
+ -  Added `IStore.PutTxIdBlockHashIndex(Guid, TxId, BlockHash)` method.
+    [[#1294], [#1328]]
+ -  Added `IStore.HasTxIdBlockHashIndex(Guid, TxId)` method.
+    [[#1294], [#1328]]
+ -  Added `IStore.GetTxIdBlockHashIndex(Guid, TxId)` method.
+    [[#1294], [#1328]]
+ -  Added `IStore.DeleteTxIdBlockHashIndex(Guid, TxId)` method.
+    [[#1294], [#1328]]
 
 ### Backward-incompatible network protocol changes
 
@@ -278,10 +286,12 @@ To be released.
 [#1285]: https://github.com/planetarium/libplanet/issues/1285
 [#1287]: https://github.com/planetarium/libplanet/pull/1287
 [#1289]: https://github.com/planetarium/libplanet/pull/1289
+[#1294]: https://github.com/planetarium/libplanet/issues/1294
 [#1298]: https://github.com/planetarium/libplanet/pull/1298
 [#1301]: https://github.com/planetarium/libplanet/issues/1301
 [#1305]: https://github.com/planetarium/libplanet/pull/1305
 [#1325]: https://github.com/planetarium/libplanet/pull/1325
+[#1328]: https://github.com/planetarium/libplanet/pull/1328
 
 
 Version 0.11.1